Добавлены команды и улучшена структура проекта
В проекте `OzonEdu.MerchandiseService.Application.csproj` добавлены ссылки на пакеты `MediatR` и проект `OzonEdu.MerchandiseService.Domain`, а также удалены некоторые папки.

Удалены старые определения класса `MerchandiseRequestGaveOut`, замененные на `MerchandiseRequestGiveOut`. Изменен тип возвращаемого значения метода `CreateAsync` в интерфейсе `IMerchandiseRepository` с `Task<int>` на `Task<long>`.

В классе `MerchandiseRequest` улучшено именование параметров и свойств, а также изменена логика метода `GiveOut` для использования нового имени свойства `GiveOutAt`.

Введены новые команды `GiveOutMerchandiseCommand` и `NewMerchandiseAppearedCommand`, а также добавлены интерфейсы `IEmailService` и `IStockApiIntegration`.

Реализована логика обработки команды выдачи мерча в классе `GiveOutMerchandiseCommandHandler` и добавлены свойства для события выдачи мерча в классе `MerchandiseRequestGiveOut`.

+++ OzonEdu.MerchandiseService.Domain/AggregationModels/MerchandiseRequests/Interfaces/IMerchandiseRepository.cs
@@ -8,7 +8,7 @@
 		/// <param name="merchandiseRequest"></param>
 		/// <param name="cancellationToken"></param>
 		/// <returns></returns>
-		Task<int> CreateAsync(MerchandiseRequest merchandiseRequest, CancellationToken cancellationToken);
+		Task<long> CreateAsync(MerchandiseRequest merchandiseRequest, CancellationToken cancellationToken);
 
 		/// <summary>
 		/// Обновление заявки на выдачу мерча
